Change errorCode to int for consistency between 32-bit and 64-bit platforms
Unfortunately this requires updating all client code that uses these callbacks.

+++ src/Connection.c
@@ -5,7 +5,7 @@ static int stage = STAGE_NONE;
 static ConnListenerConnectionTerminated originalTerminationCallback;
 static int alreadyTerminated;
 static PLT_THREAD terminationCallbackThread;
-static long terminationCallbackErrorCode;
+static int terminationCallbackErrorCode;
 
 // Common globals
 char* RemoteAddrString;
@@ -141,7 +141,7 @@ static void terminationCallbackThreadFunc(void* context)
 // calls LiStopConnection() in the callback when the cleanup code
 // attempts to join the thread that the termination callback (and LiStopConnection)
 // is running on.
-static void ClInternalConnectionTerminated(long errorCode)
+static void ClInternalConnectionTerminated(int errorCode)
 {
     int err;
